+++ xsolla-payments-sdk/src/main/java/com/xsolla/android/payments/callbacks/GetStatusCallback.kt
@@ -1,0 +1,8 @@
+package com.xsolla.android.payments.callbacks
+
+import com.xsolla.android.payments.entity.response.InvoicesDataResponse
+
+interface GetStatusCallback {
+    fun onSuccess(data: InvoicesDataResponse?)
+    fun onError(throwable: Throwable?, errorMessage: String?)
+}