+++ xsolla-payments-sdk/src/main/java/com/xsolla/android/payments/callbacks/StatusReceivedCallback.kt
@@ -1,0 +1,7 @@
+package com.xsolla.android.payments.callbacks
+
+import com.xsolla.android.payments.entity.response.InvoicesDataResponse
+
+interface StatusReceivedCallback {
+    fun onSuccess(data: InvoicesDataResponse)
+}