+++ xsolla-payments-sdk/src/main/java/com/xsolla/android/payments/ui/ActivityPayStation.kt
@@ -5,6 +5,7 @@ import android.annotation.SuppressLint
 import android.app.Activity
 import android.app.DownloadManager
 import android.content.ActivityNotFoundException
+import android.content.Context
 import android.content.Intent
 import android.content.pm.PackageManager
 import android.net.Uri
@@ -38,6 +39,7 @@ internal class ActivityPayStation : AppCompatActivity() {
         const val ARG_URL = "url"
         const val ARG_REDIRECT_SCHEME = "redirect_scheme"
         const val ARG_REDIRECT_HOST = "redirect_host"
+        const val ARG_PAYMENT_TOKEN = "payment_token"
 
         /**
          * An intent parameter used to specify activity [ActivityType].
@@ -81,6 +83,10 @@ internal class ActivityPayStation : AppCompatActivity() {
          * splash screen to fade-out into the actual PayStation content.
          */
         private const val TRUSTED_WEB_ACTIVITY_FADE_OUT_TIME_IN_MILLIS = 250
+
+        fun checkAvailability(context: Context) =
+            BrowserUtils.isPlainBrowserAvailable(context)
+                    || BrowserUtils.isCustomTabsBrowserAvailable(context)
     }
 
     private lateinit var url: String
@@ -90,6 +96,7 @@ internal class ActivityPayStation : AppCompatActivity() {
 
     private lateinit var redirectScheme: String
     private lateinit var redirectHost: String
+    private var paymentToken: String? = null
     private lateinit var type: ActivityType
     private var orientationLock: ActivityOrientationLock? = null
     private var trustedWebActivityBackgroundColor: Int? = null
@@ -116,6 +123,10 @@ internal class ActivityPayStation : AppCompatActivity() {
 
         redirectScheme = intent.getStringExtra(ARG_REDIRECT_SCHEME)!!
         redirectHost = intent.getStringExtra(ARG_REDIRECT_HOST)!!
+        paymentToken = null
+        intent.getStringExtra(ARG_PAYMENT_TOKEN)?.let { token ->
+            paymentToken = token
+        }
 
         type = intent.getStringExtra(ARG_ACTIVITY_TYPE)
             ?.let { s -> ActivityType.valueOf(s.uppercase()) }
@@ -123,6 +134,11 @@ internal class ActivityPayStation : AppCompatActivity() {
             // to the "deprecated" method, i.e. via [ARG_USE_WEBVIEW]
             // intent parameter.
             ?: if (BrowserUtils.isCustomTabsBrowserAvailable(this)) ActivityType.CUSTOM_TABS else ActivityType.WEB_VIEW
+
+
+        if(!checkAvailability(this)) {
+            type = ActivityType.WEB_VIEW
+        }
 
         orientationLock = intent.getStringExtra(ARG_ACTIVITY_ORIENTATION_LOCK)
             ?.let { s -> ActivityOrientationLock.valueOf(s.uppercase()) }
@@ -179,7 +195,8 @@ internal class ActivityPayStation : AppCompatActivity() {
             if (!isWebView()) {
                 finishWithResult(
                     Activity.RESULT_CANCELED,
-                    XPayments.Result(XPayments.Status.CANCELLED, null)
+                    XPayments.Result(XPayments.Status.CANCELLED, null),
+                    true
                 )
             }
         }
@@ -202,7 +219,8 @@ internal class ActivityPayStation : AppCompatActivity() {
 
         finishWithResult(
             Activity.RESULT_OK,
-            XPayments.Result(status, invoiceId)
+            XPayments.Result(status, invoiceId),
+            status != XPayments.Status.COMPLETED
         )
     }
 
@@ -257,8 +275,8 @@ internal class ActivityPayStation : AppCompatActivity() {
 
             override fun doUpdateVisitedHistory(view: WebView, url: String, isReload: Boolean) {
                 val uri = Uri.parse(url)
-                if (uri.scheme == redirectScheme
-                    && uri.host == redirectHost
+                if (uri.scheme.equals(redirectScheme, true)
+                    && uri.host.equals(redirectHost, true)
                 ) {
                     val invoiceId = uri.getQueryParameter("invoice_id")
                     val statusParam = uri.getQueryParameter("status")
@@ -272,7 +290,8 @@ internal class ActivityPayStation : AppCompatActivity() {
                     webView.visibility = View.INVISIBLE
                     finishWithResult(
                         Activity.RESULT_OK,
-                        XPayments.Result(status, invoiceId)
+                        XPayments.Result(status, invoiceId),
+                        status != XPayments.Status.COMPLETED
                     )
                 }
                 super.doUpdateVisitedHistory(view, url, isReload)
@@ -337,10 +356,13 @@ internal class ActivityPayStation : AppCompatActivity() {
         }
     }
 
-    private fun finishWithResult(resultCode: Int, resultData: XPayments.Result) {
+    private fun finishWithResult(resultCode: Int, resultData: XPayments.Result, isManually: Boolean) {
         val intent = Intent()
         intent.putExtra(RESULT, resultData)
         setResult(resultCode, intent)
+        paymentToken?.let { token ->
+            XPayments.payStationWasClosed(token, isManually)
+        }
         finish()
     }
 
@@ -364,6 +386,17 @@ internal class ActivityPayStation : AppCompatActivity() {
         super.onDestroy()
 
         TrustedWebActivity.notifyOnDestroy()
+    }
+
+    override fun onBackPressed() {
+        super.onBackPressed()
+        if(isWebView()) {
+            finishWithResult(
+                Activity.RESULT_CANCELED,
+                XPayments.Result(XPayments.Status.CANCELLED, null),
+                true
+            )
+        }
     }
 
     override fun onEnterAnimationComplete() {